refactor: split llm_optimization.py into utils/llm/ subpackage (Phase 2.3)
The 1,675-LOC `maverick_mcp/utils/llm_optimization.py` is now a thin
re-export shim. The implementation lives in 6 focused submodules
under `maverick_mcp/utils/llm/`, one class family per module:

| New module | Classes |
|------------|---------|
| `model_selection.py` | `ModelConfiguration`, `AdaptiveModelSelector` |
| `token_budgeter.py` | `ResearchPhase`, `TokenAllocation`, `ProgressiveTokenBudgeter` |
| `parallel.py` | `ParallelLLMProcessor` |
| `prompts.py` | `OptimizedPromptEngine` |
| `confidence.py` | `ConfidenceTracker` |
| `content_filter.py` | `IntelligentContentFilter` |

`maverick_mcp/utils/llm/__init__.py` re-exports the full public API
so callers can use either path:

    from maverick_mcp.utils.llm import AdaptiveModelSelector  # new
    from maverick_mcp.utils.llm_optimization import AdaptiveModelSelector  # legacy

The legacy `llm_optimization.py` is now a one-statement shim that
forwards to the subpackage. No production code changes — just file
moves.

Behavior is verified by the characterization tests added in PR #164
(19 cases covering the previously under-tested classes plus a public
API smoke test) plus the existing
`tests/test_optimized_research_agent.py` (16 cases) and
`tests/test_speed_optimization_validation.py` (3 collected).
All 35 llm-related tests pass against the new split.

Result: 885 unit tests pass; strict-zone ty check still 0
diagnostics; ruff clean.
